Fix OpenCV flavors
Only one OpenCV flavor can be installed in the environment, because they
share the same namespace (cv2). If OpenCV is already installed, use
that flavor, otherwise use the headless flavor (opencv-python-headless).

It doesn't seem to be possible to have this behaviour with a simple
pyproject.toml file, so dependencies are now specified in the setup.py
file instead.